refactor(alias): $EDITOR set properly, new e alias
`$EDITOR` should block the terminal, so programs invoking it won't keep
running when it should be waiting for a file. `$EDITOR` also opens a new
Emacs frame.

`e` alias now directly calls emacsclient. This is so I can decide if I
want a new frame, or using an existing one.